DCS-1348 fixing unit tests and restructuring spec folders
