+++ server/routes/bookedtoday/arrivals/courtreturns/confirmCourtReturnController.test.ts
@@ -1,12 +1,12 @@
 import type { Express } from 'express'
 import request from 'supertest'
 import cheerio from 'cheerio'
-import { appWithAllRoutes, flashProvider } from '../../__testutils/appSetup'
-import PrisonService from '../../../services/prisonService'
-import Role from '../../../authentication/role'
-import config from '../../../config'
+import { appWithAllRoutes, flashProvider } from '../../../__testutils/appSetup'
+import PrisonService from '../../../../services/prisonService'
+import Role from '../../../../authentication/role'
+import config from '../../../../config'
 
-jest.mock('../../../services/prisonService')
+jest.mock('../../../../services/prisonService')
 const prisonService = new PrisonService(null, null) as jest.Mocked<PrisonService>
 let app: Express
