+++ server/routes/bookedtoday/arrivals/courtreturns/confirmCourtReturnController.ts
@@ -1,5 +1,5 @@
 import { RequestHandler } from 'express'
-import type { PrisonService } from '../../../services'
+import type { PrisonService } from '../../../../services'
 
 export default class ConfirmCourtReturnAddedToRollController {
   public constructor(private readonly prisonService: PrisonService) {}
@@ -14,7 +14,7 @@ export default class ConfirmCourtReturnAddedToRollController {
 
       const prison = await this.prisonService.getPrison(activeCaseLoadId)
 
-      return res.render('pages/bookedtoday/arrivals/confirmCourtReturnAddedToRoll.njk', {
+      return res.render('pages/bookedtoday/arrivals/courtreturns/confirmCourtReturnAddedToRoll.njk', {
         firstName,
         lastName,
         prisonNumber,
