+++ server/views/pages/bookedtoday/arrivals/courtreturns/checkCourtReturn.njk
@@ -1,4 +1,4 @@
-{% extends "../../../partials/layout.njk" %}
+{% extends "../../../../partials/layout.njk" %}
 {% set pageTitle = "This person is returning from court" %}
 
 {% block content %}
