+++ server/views/pages/bookedtoday/arrivals/courtreturns/confirmCourtReturnAddedToRoll.njk
@@ -1,4 +1,4 @@
-{% extends "../../../partials/layout.njk" %}
+{% extends "../../../../partials/layout.njk" %}
 {% set pageTitle = "This person has returned to prison" %}
 
 {% block content %}
